Fix #2: Can't compile json gem on stock setup
Instead of using the lower-level json gem which requires ruby.h and
possibly more, just require json/pure with no external dependencies. It
may not be as performant as the json gem, but this is hardly measurable
compared to the time it takes to just retrieve the data via the
appFigures API over the net.